Refactor: 홈 화면 이동 라우터 변경
