+++ src/shared/const/navButtonUrls.ts
@@ -1,5 +1,5 @@
 export const NAV_ITEMS = [
-  { key: 'home', url: '/' },
+  { key: 'home', url: '/home' },
   { key: 'career', url: 'career' },
   { key: 'skills', url: 'skills' },
   { key: 'contact', url: 'contact' },
